requirements.txt
requests~=2.25.1
Pillow~=8.1.2
python-dotenv~=0.16.0
